Split file CNC to file GCode.
Improved the adding to the path variable, such that always the local
files get preferred.
